# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SabayTenh eCommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  cashier: 'cashier',
  staff: 'staff',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  customer: 'customer',
  cashier: 'cashier',
  staff: 'staff',
  admin: 'admin',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  nameKh?: string | null;
  slug: string;
  /** @nullable */
  icon?: string | null;
  productCount?: number;
}

export interface Brand {
  id: number;
  name: string;
  /** @nullable */
  logo?: string | null;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  nameKh?: string | null;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  /** @nullable */
  image?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  brandName?: string | null;
  stock: number;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  isFeatured?: boolean;
  /** @nullable */
  tags?: string | null;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt?: string;
}

export interface ProductDetail {
  id: number;
  name: string;
  /** @nullable */
  nameKh?: string | null;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discountPercent?: number | null;
  images?: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  brandName?: string | null;
  stock: number;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  isFeatured?: boolean;
  /** @nullable */
  tags?: string | null;
  reviews?: Review[];
}

export interface ProductInput {
  name: string;
  nameKh?: string;
  description?: string;
  price: number;
  originalPrice?: number;
  image?: string;
  categoryId: number;
  brandId?: number;
  stock: number;
  isFeatured?: boolean;
  tags?: string;
}

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface ReviewInput {
  rating: number;
  comment?: string;
}

export interface CartItem {
  id: number;
  productId: number;
  productName?: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
  subtotal?: number;
}

export interface Cart {
  id: number;
  items: CartItem[];
  subtotal: number;
  discount?: number;
  total: number;
  itemCount?: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  subtotal?: number;
  discount?: number;
  total: number;
  /** @nullable */
  paymentMethod?: string | null;
  paymentStatus?: OrderPaymentStatus;
  /** @nullable */
  shippingAddress?: string | null;
  itemCount?: number;
  createdAt: string;
}

export type OrderDetailStatus = typeof OrderDetailStatus[keyof typeof OrderDetailStatus];


export const OrderDetailStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderDetailPaymentStatus = typeof OrderDetailPaymentStatus[keyof typeof OrderDetailPaymentStatus];


export const OrderDetailPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  productName?: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  price: number;
  subtotal?: number;
}

export interface OrderDetail {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderDetailStatus;
  subtotal?: number;
  discount?: number;
  total: number;
  /** @nullable */
  paymentMethod?: string | null;
  paymentStatus?: OrderDetailPaymentStatus;
  /** @nullable */
  shippingAddress?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cash: 'cash',
  khqr: 'khqr',
  aba: 'aba',
  acleda: 'acleda',
  canadia: 'canadia',
  wing: 'wing',
} as const;

export interface OrderInput {
  paymentMethod: OrderInputPaymentMethod;
  shippingAddress: string;
  note?: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface InventoryItem {
  productId: number;
  productName: string;
  /** @nullable */
  categoryName?: string | null;
  stock: number;
  lowStockThreshold: number;
  isLowStock?: boolean;
  price?: number;
}

export interface InventoryUpdate {
  stock: number;
  lowStockThreshold?: number;
}

export interface CategorySales {
  categoryName: string;
  revenue: number;
  orderCount: number;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  todayRevenue: number;
  todayOrders: number;
  lowStockCount: number;
  pendingOrders: number;
  topCategories?: CategorySales[];
  bestSellers?: Product[];
}

export interface DailySales {
  date: string;
  revenue: number;
  orders: number;
}

export type ListProductsParams = {
search?: string;
categoryId?: number;
brandId?: number;
minPrice?: number;
maxPrice?: number;
minRating?: number;
page?: number;
limit?: number;
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating_desc: 'rating_desc',
  newest: 'newest',
  best_selling: 'best_selling',
} as const;

export type ListOrdersParams = {
status?: string;
page?: number;
};

export type ListInventoryParams = {
lowStock?: boolean;
};