# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SabayTenh eCommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (customer sees own, staff sees all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (staff/cashier)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: Get inventory overview
      parameters:
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /inventory/{productId}:
    put:
      operationId: updateInventory
      tags: [inventory]
      summary: Update stock for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, cashier, staff, admin]
        createdAt:
          type: string
...
    Cart:
      type: object
      required: [id, items, subtotal, total]
      properties:
        id:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        itemCount:
          type: integer
    CartItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    Order:
      type: object
      required: [id, userId, status, total, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        shippingAddress:
          type: ["string", "null"]
        itemCount:
          type: integer
        createdAt:
          type: string
    OrderDetail:
      type: object
      required: [id, userId, status, total, items, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        shippingAddress:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    OrderInput:
      type: object
      required: [paymentMethod, shippingAddress]
      properties:
        paymentMethod:
          type: string
          enum: [cash, khqr, aba, acleda, canadia, wing]
        shippingAddress:
          type: string
        note:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
    InventoryItem:
      type: object
      required: [productId, productName, stock, lowStockThreshold]
      properties:
        productId:
          type: integer
        productName:
          type: string
        categoryName:
          type: ["string", "null"]
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        price:
          type: number
    InventoryUpdate:
      type: object
      required: [stock]
      properties:
        stock:
          type: integer
        lowStockThreshold:
          type: integer
...