# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SabayTenh eCommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  subtotal?: number;
  discount?: number;
  total: number;
  /** @nullable */
  paymentMethod?: string | null;
  paymentStatus?: OrderPaymentStatus;
  /** @nullable */
  shippingAddress?: string | null;
  itemCount?: number;
  createdAt: string;
}